SBERDOMA-151 Editing of selected text in AddressSearchInput
`Select` component from Ant displays either `searchText` or selected `Option`,
which are two DOM elements on the same position. It's not supposed by Ant to display both of them.
Once `Option` is selected, `searchText` (typed by hands) will be wiped out.
Out of the box, it's impossible to continue editing what user has selected,
because `Option` is not editable, it's selectable.
To make it possible from visual point of view, we need to:
1. Set `searchText` to text of selected `Option` to let user to reuse selected Option's text and edit it
2. Clear `value`, that causes selected Option to disappear