SBERDOMA-151 replaced Select to Input and Dropdown
It's impossible to implement editing of selected option using Ant Select, because Option is not editable, it is selectable.
Instead, Input is used for entering address, and Dropdown for suggestions.
Custom validation is implemented, that checks whether typed value matches one item from suggestions.
If it's valid, then it will be followed up to `onChange` property, otherwise `null` will be followed up.