SBERDOMA-347 changeTrackable utility module
Utilities to make a GQLListSchema item trackable for changes.
We take a "source" schema, whose changes we want to track and
store actual changes in "destination" schema.

For example, to track changes of `Ticket` schema, following should be done:
1. Decide, what fields should be trackable, using `trackableFieldsFrom` function;
2. Create a "destination" schema, e.g. `TicketChange`
3. Generate fields for it using `generateChangeTrackableFieldsFrom` function
4. Write `DisplayNameResolvers` (see description below in typedef)
5. Write `RelatedManyToManyResolvers` (see description below in typedef)
6. Implement an `afterChange` hook in `Ticket` and call `afterChangeHook` inside of it.

That's it!
Changes of `Ticket` will be stored in `TicketChange` schema!