SBERDOMA-434 new domain `contact`, TicketContact → Contact
New domain `contact` has been created and `TicketContact` schema has been moved to `Contact`.

From a real-world perspective, a `Contact` — is a person, that is handled by a business workflows.
When the `Contact` will be tied to context of a `Ticket` only, it will cause following problems in near future:

1. It will be quickly get unclear, why we have a `TicketContact`, when there is no ticket, associated with the contact item in case of new, different workflow;
2. It will cause duplication and association problems, when new `…Contact` schemas will be introduced;
3. A `ticket` domain will quickly be bloated, when new features for contacts will be implemented (e.g. import, export).

To avoid this problems, to reduce refactoring time costs, we have decided to keep a general-purpose `Contact` item.