SBERDOMA-1057 replaced `property` to `address` in check on duplicates on create Resident
Previously, resident was checked on duplicates using a set of `property`, `unitName` and `user` fields.
When we are trying to create a resident on different address, but with same unitName (both without properties), false-positive duplicates check error will be returned.

Logically, no resident of the same user can exist in the same address and unit name (flat).

Therefore, check for the same property (of organization) is not needed.