DOMA-1215 changed logic of `syncSubscriptions` function to handle data changes, not all data
According to official comments from Fintech API representatives, `/v1/partner-info/advance-acceptances` method returns data **changes** only for current day.
It does not just returns all subscriptions for services of our organization, It's even not possible ;)

So, when nothing is returned, do nothing.
When some subscriptions are returned, check, whether it is active or not.