DOMA-3099 replaced entire Ant notification content with custom `TasksProgress` component
We need to control visual state of component dynamically, including icon (in future) and button.
Ant does not provides this possibility, it allows only to redisplay notification with the same key.
So, I've utilized only `notification({ description })` argument and placed entire `TasksProgress` component, that is looking the same as normal full-featured notification.