DOMA-3825 display only today tasks
Prevents stucking with tasks, not completed for some reason, created in previous days

Open question about pivot point today/tomorrow, that arises an answer — why not to display tasks only in 24 hours.
From working schedule point of view it's better to relate on calendar rather then durations, so, an employee opens a page on the beginning of the day and he dont see stuck tasks from yesterday.