DOMA-3778 fixed wrong translation in `exportTicketTask`
A problem was in different locales used to translate `TicketStatus.name`, becaues of different execution contexts.
A dictionary of statuses `statusNames` that was passed to `_buildExportFile` template rendering function has used locale from `TicketExportTask`, which is a proper way, because it is a request of a user. But values from `TicketStatus.name` field of type `LocalizedText` vere trying to use locale of `context.req` which was undefined, because a GraphQL query was executed in worker, not in scope of HTTP-request. Therefore a locale was undefined and a fallback to `DEFAULT_LOCALE` was used.

So, I've created a new utility `setLocaleForKeystoneContext` for this kind of use cases which sets specified locale to execute GraphQL queries server-side without having an HTTP-request scope.