feat(condo): DOMA-4395 added new field `property` to BankAccount
NOT NULL constraint cannot be set right when adding new column "property" to "BankAccount", because there is no way to initialize it with default value that points to actual records from "Property".
An error "cannot use subquery in DEFAULT expression" will be thrown in case of using SQL-queries for "DEFAULT" part.
We don't have any records yet in BankAccount table.
So, just to make this migration to succeed, we do following:
1. Add table without NOT NULL constraint
2. Set it to "id" column value of first Property
3. Set default constraint